+++ samples/net/updatehub/CMakeLists.txt
@@ -10,6 +10,7 @@ set(KCONFIG_ROOT ${CMAKE_CURRENT_SOURCE_DIR}/Kconfig)
 include($ENV{ZEPHYR_BASE}/cmake/app/boilerplate.cmake NO_POLICY_SCOPE)
 project(NONE)
 
+target_include_directories(app PRIVATE $ENV{ZEPHYR_BASE}/lib/updatehub/include)
 target_sources(app PRIVATE src/main.c)
 
 if(CONFIG_UPDATEHUB_DTLS)
